+++ include/update/updater.h
@@ -25,7 +25,7 @@
 
 #include <mutex>
 #include <string>
-#include "network/curleasy.h"
+#include <cpr/callback.h>
 #include "version.h"
 
 namespace Nickvision::Update
@@ -64,10 +64,10 @@ namespace Nickvision::Update
          * @brief Downloads and installs an application update for Windows. 
          * @brief fetchCurrentVersion should be called first before running this method.
          * @param versionType The type of version update to install
-         * @param progress An optional CurlProgressFunction to track the downloading update
+         * @param progress An optional cpr::ProgressCallback to track the downloading update
          * @return True if successful, else false
          */
-        bool windowsUpdate(VersionType versionType, const Network::CurlProgressFunction& progress = {});
+        bool windowsUpdate(VersionType versionType, const cpr::ProgressCallback& progress = {});
 #endif
         /**
          * @brief Copies an Updater.