+++ tests/webtests.cpp
@@ -30,7 +30,7 @@ TEST(WebTests, FetchJsonString1)
         GTEST_SKIP();
     }
 #endif
-    boost::json::value json = Web::fetchJson("https://api.github.com/repos/nickvisionapps/denaro/tags");
+    boost::json::value json = Web::getJson("https://api.github.com/repos/nickvisionapps/denaro/tags");
     ASSERT_FALSE(json.is_null());
     ASSERT_TRUE(json.is_array());
 }